+++ tests/Cassowary.Tests/PartialConstraintTest.cs
@@ -12,17 +12,17 @@ public class PartialConstraintTest
     public void OrWithValue()
     {
         var partial = new PartialConstraint(new Expression(
-                ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<float>())),
-                _fixture.Create<float>()),
+                ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<double>())),
+                _fixture.Create<double>()),
             _fixture.Create<WeightedRelation>());
 
-        var value = _fixture.Create<float>();
+        var value = _fixture.Create<double>();
 
         var constraint = partial | value;
         constraint.Expression.Constant.Should().Be(partial.Expression.Constant - value);
         constraint.Expression.Terms.Should().BeEquivalentTo(partial.Expression.Terms);
 
-        constraint = partial | (double)value;
+        constraint = partial | (float)value;
         constraint.Expression.Constant.Should().Be(partial.Expression.Constant - value);
         constraint.Expression.Terms.Should().BeEquivalentTo(partial.Expression.Terms);
 
@@ -30,7 +30,7 @@ public class PartialConstraintTest
         constraint.Expression.Constant.Should().Be(partial.Expression.Constant - value);
         constraint.Expression.Terms.Should().BeEquivalentTo(partial.Expression.Terms);
 
-        constraint = (double)value | partial;
+        constraint = (float)value | partial;
         constraint.Expression.Constant.Should().Be(partial.Expression.Constant - value);
         constraint.Expression.Terms.Should().BeEquivalentTo(partial.Expression.Terms);
     }
@@ -39,8 +39,8 @@ public class PartialConstraintTest
     public void OrWithVariable()
     {
         var partial = new PartialConstraint(new Expression(
-                ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<float>())),
-                _fixture.Create<float>()),
+                ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<double>())),
+                _fixture.Create<double>()),
             _fixture.Create<WeightedRelation>());
 
         var variable = new Variable();
@@ -54,13 +54,13 @@ public class PartialConstraintTest
     public void OrWithExpression()
     {
         var partial = new PartialConstraint(new Expression(
-                ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<float>())),
-                _fixture.Create<float>()),
+                ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<double>())),
+                _fixture.Create<double>()),
             _fixture.Create<WeightedRelation>());
 
         var expression = new Expression(
-            ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<float>())),
-            _fixture.Create<float>());
+            ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<double>())),
+            _fixture.Create<double>());
 
         var constraint = partial | expression;
         constraint.Expression.Constant.Should().Be(partial.Expression.Constant - expression.Constant);
@@ -71,11 +71,11 @@ public class PartialConstraintTest
     public void OrWithTerm()
     {
         var partial = new PartialConstraint(new Expression(
-                ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<float>())),
-                _fixture.Create<float>()),
+                ImmutableArray<Term>.Empty.Add(new Term(new Variable(), _fixture.Create<double>())),
+                _fixture.Create<double>()),
             _fixture.Create<WeightedRelation>());
 
-        var term = new Term(new Variable(), _fixture.Create<float>());
+        var term = new Term(new Variable(), _fixture.Create<double>());
 
         var constraint = partial | term;
         constraint.Expression.Constant.Should().Be(partial.Expression.Constant);
